requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
frozendict==2.4.6
h11==0.14.0
idna==3.10
multitasking==0.0.11
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
selenium==4.20.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
webdriver-manager==4.0.2
wsproto==1.2.0
yfinance==0.2.55
